Split honey_badger into submodules.
